Resturcture the FileManager a bit so that more functions are held within the class object. Add Copy Feature. Not fully functioning.

+++ FindingFiles.py
@@ -2,6 +2,7 @@ import os
 import glob
 import pytz
 import datetime
+import time
 import calendar
 import pywintypes
 import shutil
@@ -16,23 +17,77 @@ videoExtensions = ["*.mp4", "*.mov", "*.MP4", "*.avi", "*.mkv", "*.m4v"]
 photoExtensions = ["*.jpg", "*.heic", "*.ARW", "*.png", "*.dng", "*.jpeg"]
 pickleExtensions = ["*.pickle"]
 Destructive = False
-targetDirectory = 'E:/SSD/Media/OneSecond'
-sourceDirectory = 'E:/SSD/Media/OneSecond/oneplusVideos'
+largeDateNumberLimit = 10
+targetDirectory = 'D:/OneSecond/Testing/Destination'
+sourceDirectory = 'D:/OneSecond/Testing/Source'
 dictionaryFilename = 'dictionary.pickle'
 
-class FileInformation:
-    def __init__(self, lastDirFound, name, size):
-        self.LastDirectoryFound = lastDirFound
-        self.Name = name
-        self.Size = size
 
-class BigFileFolderInformation:
-    def __init__(self, size, files):
-        self.Size = size
-        self.Files = files
-    
+class FileInformation:
+    def __init__(self, lastDirFound):
+        self.LastDirectoryFound = lastDirFound
+        self.ExistsInCurrentDir = True
+
+        # Extract name
+        head_tail = os.path.split(lastDirFound)
+        self.Name = head_tail[1]
+
+        # Get Creation Date
+        dt = GetCreationDateFromVideo(lastDirFound)
+        self.DateTime = dt
+
+        self.Size = os.path.getsize(lastDirFound)
+        # Could add other relevent things like metadata or tags
+
+# Used when a cluster of files is large so that we can leave the files else where, but still track essential information about them
+
+
+class FileClusterManager:
+    def __init__(self):
+        self.Size = 0
+        # List of FileInformation Classes
+        self.Files = []
+        self.FilesExistOnDisk = False
+        self.Path = None
+
+    # Add files to this cluster
     def Add(self, file):
         self.Files.append(file)
+
+        # Extract directory
+        head_tail = os.path.split(file.LastDirectoryFound)
+
+        self.Path = os.path.join(head_tail[0], str(
+            file.DateTime.year), str(file.DateTime.month))
+        self.FilesExistOnDisk = True
+
+        # Here we should make the call to move the file to the cluster
+        # to create a folder and move existing files
+        return
+
+    # How many files in this cluster
+    def Number(self):
+        return len(self.Files)
+
+    def GetFiles(self):
+        return self.Files
+
+    # Return the Path that files for this cluster should go to
+    def GetPath(self):
+        return self.Path
+
+    # TODO
+    # Finding Existing files in cluster and moving them to a
+    # new folder because there are now too many to just free roam
+    def IsSecondFile(self):
+        # move existing file into this new date folder
+        for files in self.GetFiles(dt):
+            newFile = os.path.join(pathDestination, files.name)
+            Path.rename(files, newFile)
+
+            # update file in the dictionary
+            self.DeleteEntry(dt)
+            self.Add(newFile, dt)
 
 
 def ConvertToDateTime(pywintime):
@@ -45,15 +100,6 @@ def ConvertToDateTime(pywintime):
         second=pywintime.second
     )
     return now_datetime
-
-
-def FindItemsInDirectory(sourceDir, extensionList, recursive=True):
-    pathList = []
-    for ext in extensionList:
-        if recursive:
-            ext = "**/" + ext
-        pathList.extend(Path(sourceDir).glob(ext))
-    return pathList
 
 
 def MakeFolder(path):
@@ -110,39 +156,46 @@ def MakeFoldersForMonths(dir, earliestDate, latestDate):
         year = year + 1
 
 
-class Files:
+class FileManager:
 
     def __init__(self):
-        # dictionary of file locations by date
-        self.files_by_date = defaultdict(list)
+        # dictionary of file cluster managers by date
+        self.files_by_date = defaultdict(FileClusterManager)
 
         self.firstDate = datetime.date(2020, 6, 15)
         self.lastDate = datetime.date(1987, 6, 15)
 
     # Add file directory to one of the date keys
     # TODO prevent a double add
-        # probably need to document the name of the file?
-    def Add(self, file, dt):
+    # Add new FileInformation object to the dictionary
+    def Add(self, fileInfo):
+        self.files_by_date[fileInfo.DateTime.date()].Add(fileInfo)
+
         # Update timeframe
-        if dt.date() > self.lastDate:
-            self.lastDate = dt.date()
-        if dt.date() < self.firstDate:
-            self.firstDate = dt.date()
+        if fileInfo.DateTime.date() > self.lastDate:
+            self.lastDate = fileInfo.DateTime.date()
+        if fileInfo.DateTime.date() < self.firstDate:
+            self.firstDate = fileInfo.DateTime.date()
 
-        self.files_by_date[dt.date()].append(file)
-
-    # Delete the entire list entry for the date key
+    # Delete the entire FileInformationManager entry for the date key
     def DeleteEntry(self, dt):
         del self.files_by_date[dt.date()]
 
-    # Get the list of file directories on the date key
+    def FindItemsInDirectory(self, srcDir, extensionList, recursive=True):
+        pathList = []
+        for ext in extensionList:
+            if recursive:
+                ext = "**/" + ext
+            pathList.extend(Path(srcDir).glob(ext))
+        return pathList
+
+    # Get the list of file Information objects on the date key
     def GetFiles(self, date):
-        return self.files_by_date[date]
+        return self.files_by_date[date].GetFiles()
 
     # Get the number of file directores on a date key
     def NumberOfFilesOnDate(self, date):
-        obj = self.files_by_date[date]
-        return len(obj)
+        self.files_by_date[date].GetNumber()
 
     # Get the oldest date key in the dictionary
     def OldestFile(self):
@@ -222,10 +275,11 @@ class Files:
     def GetNumberOfDaysThatHaveFilesInAMonth(self, month, year):
         runningCount = 0
         for dateKey in self.files_by_date:
-            if (dateKey.year == year and dateKey.month == month and len(self.files_by_date[dateKey]) > 0):
+            if (dateKey.year == year and dateKey.month == month and (self.files_by_date[dateKey].Number()) > 0):
                 runningCount = runningCount + 1
         return runningCount
 
+    # TODO
     def TripDetector(self):
         # Analyze missing days and see if there would be available clips in the surrounding days
         return
@@ -238,12 +292,24 @@ class Files:
             if entry < self.firstDate:
                 self.firstDate = entry
 
-    def SaveDictionary(self, directory, fileName):
+    def SavePickleFile(self, directory, fileName):
         fileDir = os.path.join(directory, fileName)
-        with open(fileDir, 'wb') as handle:
-            pickle.dump(self.files_by_date, handle,
-                        protocol=pickle.HIGHEST_PROTOCOL)
+        if(os.path.exists(fileDir)):
+            with open(fileDir, 'wb') as handle:
+                pickle.dump(self.files_by_date, handle,
+                            protocol=pickle.HIGHEST_PROTOCOL)
         return
+
+    def LoadPickleFile(self, directory, fileName, testFunc):
+
+        fileDir = os.path.join(directory, fileName)
+
+        with open(fileDir, 'rb') as handle:
+            self.files_by_date = pickle.load(handle)
+
+        result = testFunc
+
+        return result
 
     def TestDictionary(self, depth=5):
         # Depth: is how far into the dictionary we should search before validating it
@@ -254,103 +320,161 @@ class Files:
         for entry in self.files_by_date:
             if (len(self.files_by_date[entry]) >= 0):
                 pathsChecked = pathsChecked + 1
-                for path in self.files_by_date[entry]:
-                    if(Path.exists(path)):
-                        pathsSuccessful = pathsSuccessful + 1
-                    if(pathsSuccessful >= depth):
-                        print("Test passed on found dictionary")
-                        return True
-                    if(pathsChecked > depth):
-                        return False
+                for fileInfo in self.files_by_date[entry].GetFiles():
+                    # Make sure that we actually moved the file onto the disk
+                    if fileInfo.ExistsInCurrentDir:
+                        path = fileInfo.File
+                        if(Path.exists(path)):
+                            pathsSuccessful = pathsSuccessful + 1
+                        if(pathsSuccessful >= depth):
+                            print("Test passed on found dictionary")
+                            return True
+                        if(pathsChecked > depth):
+                            return False
+
+    def SaveDictionary(self, directory, fileName):
+        self.SavePickleFile(directory, fileName)
 
     def LoadDictionary(self, directory, fileName):
+        return self.LoadPickleFile(directory, fileName, self.TestDictionary)
 
-        fileDir = os.path.join(directory, fileName)
+    def FindExistingDictionary(self, directory):
+        pickles = self.FindItemsInDirectory(directory, pickleExtensions, False)
 
-        with open(fileDir, 'rb') as handle:
-            self.files_by_date = pickle.load(handle)
+        if len(pickles) > 0:
+            # found an existing directory, only use the first one found
+            self.LoadDictionary(directory, pickles[0])
+            return True
 
-        result = self.TestDictionary()
-        return result
-    
+        else:
+            return False
 
+    def IndexVideosInDirectory(self, directory):
+        # find videos in the directory
+        existingVideos = self.FindItemsInDirectory(directory, videoExtensions)
+
+        # add video file to the dictionary
+        for vid in existingVideos:
+            fileInfo = FileInformation(vid)
+            self.Add(fileInfo)
+
+    def DeletePhotosInDirectory(self, directory):
+        # find files that are photos
+        photoPaths = self.FindItemsInDirectory(directory, photoExtensions)
+
+        # delete files that are photos
+        for file in photoPaths:
+            print("deleting file")
+            file.unlink()
+
+    # A blocking file copy function
+    def CopyFileToDir(self, srcDir, dstDir):
+
+        # Make folders if necessary and move files
+        if not (os.path.exists(dstDir)):
+            MakeFolder(dstDir)
+
+        # copy 2 can have dst to be a directory
+        shutil.copy2(srcDir, dstDir)
+
+        historicalSize = -1
+        while (historicalSize != os.path.getsize(dstDir)):
+            historicalSize = os.path.getsize(dstDir)
+            time.sleep(3)
+        print('File copy has now finished')
+
+    # Cut videos from an existing directory and move it to a new target dir
+    def CutVideosFromDirectory(self, sourceDir, targetDir):
+        # Should we delete videos from the sourceDir if the FileClusterManager is not accepting any more videos
+
+        # Find all the videos in the sourceDir
+        videoPaths = self.FindItemsInDirectory(sourceDir, videoExtensions)
+
+        for filePath in videoPaths:
+
+            # Create a new FileInformation object given the path
+            fileInfo = FileInformation(filePath)
+
+            # Get the destination path from the FileClusterManager
+            dstPath = self.files_by_date[fileInfo.DateTime.date()].GetPath()
+
+            # Add the file name to the end of the destination directory
+            dstPath_File = os.path.join(dstPath, fileInfo.Name)
+
+            if(self.CopyFileToDir(fileInfo.LastDirectoryFound, dstPath_File)):
+               # Confirm that the file has been copied
+                if(os.path.exists(dstPath_File)):
+                    # Copy was successfuly, update its last known directory
+                    fileInfo.LastDirectoryFound = dstPath_File
+                else:
+                    print('Failed to copy file')
+                    return False
+
+                # Add it to our dictionary
+                self.Add(fileInfo)
+
+                # Now that it is added to the dictionary, we can delete it from the old directory
+                os.remove(sourceDir)
+            else:
+                return False
+        return True
+
+    # Copy videos from an existing directory and paste it to a new target dir
+    def CopyVideosFromDirectory(self, sourceDir, targetDir):
+        # Find all the videos in the sourceDir
+        videoPaths = self.FindItemsInDirectory(sourceDir, videoExtensions)
+
+        for filePath in videoPaths:
+
+            # Create a new FileInformation object given the path
+            fileInfo = FileInformation(filePath)
+
+            # Get the destination path from the FileClusterManager
+            dstPath = self.files_by_date[fileInfo.DateTime.date()].GetPath()
+
+            # Add the file name to the end of the destination directory
+            dstPath_File = os.path.join(dstPath, fileInfo.Name)
+
+            if(self.CopyFileToDir(fileInfo.LastDirectoryFound, dstPath)):
+                # Confirm that the file has been copied
+                if(os.path.exists(dstPath_File)):
+                    # Copy was successfuly, update its last known directory
+                    fileInfo.LastDirectoryFound = dstPath_File
+                else:
+                    print('Failed to copy file')
+                    return False
+
+                # Add it to our dictionary
+                self.Add(fileInfo)
+            else:
+                return False
+        return True
 
 
 def main():
 
-    ref = Files()
+    ref = FileManager()
 
-    pickles = FindItemsInDirectory(targetDirectory, pickleExtensions, False)
+    if(ref.FindExistingDictionary(targetDirectory)):
+        print("Existing Dictionary Found")
 
-    if len(pickles) > 0:
-        # found an existing directory
-        ref.LoadDictionary(targetDirectory, pickles[0])
+        # TODO Decide if we need to update the dictionary by looking at the
+        # file modified times of the highest level folders vs the pickle file
+
     else:
-        # index the the videos that already exist in the folders
-        existingVideos = FindItemsInDirectory(targetDirectory, videoExtensions)
 
-        # add video paths to the dictionary
-        for vid in existingVideos:
-            dt = GetCreationDateFromVideo(vid)
-            ref.Add(vid, dt)
+        # Add videos to dictionary that are already in the Target Directory
+        ref.IndexVideosInDirectory(sourceDirectory)
 
         if Destructive:
-            # find files that are photos
-            photoPaths = FindItemsInDirectory(sourceDirectory, photoExtensions)
+            ref.DeletePhotosInDirectory(sourceDirectory)
 
-            # delete files that are photos
-            for file in photoPaths:
-                print("deleting file")
-                file.unlink()
+            # Add videos to dictionary and then copy them
+            ref.CutVideosFromDirectory(sourceDirectory, targetDirectory)
 
-        # digs up the matching video extensions
-        videoPaths = FindItemsInDirectory(sourceDirectory, videoExtensions)
-
-        for file in videoPaths:
-            # Get Creation Date
-            dt = GetCreationDateFromVideo(file)
-
-            # If it is the second file with that date record
-            if(ref.NumberOfFilesOnDate(dt.date()) > 1):
-                pathDestination = os.path.join(
-                    targetDirectory, str(dt.year), str(dt.month), str(dt.day))
-
-                # move existing file into this new date
-                for files in ref.GetFiles(dt):
-                    newFile = os.path.join(pathDestination, files.name)
-                    Path.rename(files, newFile)
-
-                    # update file in the dictionary
-                    ref.DeleteEntry(dt)
-                    ref.Add(newFile, dt)
-
-            else:
-                pathDestination = os.path.join(
-                    targetDirectory, str(dt.year), str(dt.month))
-
-            # Make folders if necessary and move files
-            if not (os.path.exists(pathDestination)):
-                MakeFolder(pathDestination)
-
-            # Create the new file path
-            newFile = os.path.join(pathDestination, file.name)
-
-            # Move path
-            if os.path.exists(newFile):
-                if Destructive:
-                    # remove video in the sourceDir
-                    Path.unlink(file)
-            else:
-                try:
-                    # If the file exists, move it
-                    Path.rename(file, newFile)
-                except OSError:
-                    print('Exception while trying to move file: ', newFile)
-                    print('Error: ', OSError)
-                else:
-                    print('File Moved: ', newFile)
-
-            ref.Add(file, dt)
+        else:
+            # Add videos to dictionary and then delete them
+            ref.CopyVideosFromDirectory(sourceDirectory, targetDirectory)
 
     ref.PrintNumberOfMissingDatesPerMonth(2019)
     ref.SaveDictionary(targetDirectory, dictionaryFilename)
